# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CLISENSA API
  description: API REST de la plataforma multi-hospital de citas médicas CLISENSA.
  version: 0.5.0
servers:
  - url: http://localhost:4000/api
    description: Desarrollo local
...
paths:
...
  /medicos:
    get:
      tags: [Medicos]
      summary: Listar médicos y sus horarios base
      parameters:
        - in: query
          name: especialidad
          schema: { type: string }
        - in: query
          name: hospitalId
          schema: { type: integer }
      responses:
        "200":
          description: Lista de médicos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicoConHorarios"
  /medicos/invitar:
    post:
      tags: [Medicos]
      summary: Invitar médico y enviar enlace de activación
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [licenciaMedica, correo, especialidad]
              properties:
                licenciaMedica: { type: string }
                correo: { type: string, format: email }
                especialidad: { type: string }
      responses:
        "201": { description: Invitación enviada }
        "409": { description: El correo o la licencia ya existe }
        "503": { description: Configuración de correo incompleta }
  /medicos/perfil:
    get:
      tags: [Medicos]
      summary: Consultar perfil del médico autenticado
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Perfil médico }
    put:
      tags: [Medicos]
      summary: Completar perfil, activar cuenta o actualizar perfil del médico
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombreCompleto: { type: string }
                biografia: { type: string }
                fotoUrl: { type: string, format: uri, nullable: true }
                password: { type: string, minLength: 8 }
      responses:
        "200": { description: Perfil actualizado }
        "401": { description: Token de activación inválido o vencido }
    post:
      tags: [Medicos]
      summary: Crear médico (rol administrativo)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, especialidad]
              properties:
                nombre: { type: string, example: Dra. L. Torres }
                especialidad: { type: string, example: Ginecologia }
      responses:
        "201":
          description: Médico creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medico"
        "400": { description: Faltan campos requeridos }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol administrativo }
  /medicos/{id}/horarios:
    post:
      tags: [Medicos]
      summary: Agregar horario base a un médico (rol administrativo)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hora]
              properties:
                hora: { type: string, pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$", example: "09:00" }
      responses:
        "201":
          description: Horario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HorarioBase"
        "400": { description: La hora falta o tiene formato inválido }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol administrativo }
        "404": { description: Médico no encontrado }
        "409": { description: El horario ya existe }
  /medicos/{id}/horarios/{horarioId}:
    delete:
      tags: [Medicos]
      summary: Quitar horario base de un médico (rol administrativo)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: horarioId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Horario eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol administrativo }
        "404": { description: Horario no encontrado para ese médico }
...
  /citas/medico:
    get:
      tags: [Citas]
      summary: Consultar agenda del médico autenticado
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: fecha
          required: true
          schema: { type: string, format: date }
      responses:
        "200": { description: Agenda del médico }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol médico }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Medico:
      type: object
      required: [id, nombre, especialidad]
      properties:
        id: { type: integer }
        nombre: { type: string }
        especialidad: { type: string }
    HorarioBase:
      type: object
      required: [id, medicoId, hora]
      properties:
        id: { type: integer }
        medicoId: { type: integer }
        hora: { type: string, example: "09:00" }
    MedicoConHorarios:
      allOf:
        - $ref: "#/components/schemas/Medico"
        - type: object
          properties:
            horarios:
              type: array
              items:
                $ref: "#/components/schemas/HorarioBase"
...